ci: configuration: add per-character delay Some devices have trouble writing the CI configuration file to their tmpfs at default (high) baudrate. As a result, the written configuration files might sporadically see missing characters or lines, causing an issue with their certification tests.
The workaround in this commit is to introduce a per-character delay (instead of per-line), effectively lowering the general baudrate while writing the configuration file.
Tested and verified on the problematic platforms.

Also lower this "send_delay", since it seems to not affect the configuration step.

This workaround only adds a few seconds per configuration step.

Signed-off-by: Maarten De Decker <maarten.dedecker@mind.be>