controller: fix invalid iterator access
In some cases, the iterator we get when looking for the current move's
BSSID is invalid (because the container doesn't contain any matching
BSSID).

It's not clear in which conditions it happens but when it does, it
leads to the controller crashing.

To fix it, check the iterator before accessing it. If it's invalid,
move on to resetting the station stats.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>