agent: bwl: set notify_mgmt_frames when dynamically creating BSSes
We need this option to be able to get the management frames when a
station connects.

The reason why we figure out this bug only now is because the
autoconfiguration flow also sets it, and it happens to be a global
option in prplOS (while it's per-BSS in upstream hostapd - that has to
be fixed).

In any case, set it when creating a BSS dynamically to make sure we
receive management frames in all cases.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>