agent: fix AP metrics memory leak
It is observed that AP metrics reponses are keep on stacked
in beerocks_agent without sending those responses to the
controller. This scenario occurs when there are pending AP
metrics Query in beerocks_agent. Due to this, Memory leak
is happening on the beerocks_agent process.

To resolve this, previous stacked AP metrics responses are
cleared while clearing the AP metric query during timeout.

Signed-off-by: Balaji Manoharan <bmanoharan@maxlinear.com>