nbapi: use libamxrt api to make prplmesh datamodel persistent
The aim of this branch is to use libamxrt as a method of persistent storage. using this branch, the controller has the capability to store data persistently, even through a reboot. In future, it might become necessary to eliminate the UCI support if NBAPI is activated, using libamxrt in prplmesh nbapi the controller will be able to load amx plug-ins like mod-pcm-svc or mod-sahtrace.

Some tweaks and fixes were made for the same purpose like the limitations of parsing $'value' by cmake when configuring odl.in files.

The persistent reboot file is saved into /opt/prplmesh/share/odl/prplmesh.odl

Signed-off-by: Ahmed ARIF <ahmed.arif_ext@softathome.com>