Show Agent and Fronthaul states in the datamodel
The Agent manages the datamodel.
Fronthauls send their states in a proprietary CMDU over Agent UDS,
and the Agent updates the DM.

Would be nice if Agent monitored the spawned Fronthaul processes
and deleted their corresponding entries on Fronthaul stop/crash,
but this is not the case (yet?), so I expose Fronthauls PIDs in the DM,
so one could check if the DM data is actual.

Fixes PPM-2635

Signed-off-by: Andrey Erokhin <a.erokhin@inango-systems.com>