tlvf_utils: Speed-up AP Capability Report creation
To prepare the AP Capability Report TLV, the EasyMesh Agent has to go through
each channel and bandwidth to determine whether they are operable or not.
This process can take up to several 100s of milliseconds. The most time-consuming
operation is the call to get_center_channel().
Currently, this call is performed for each tuple (channel, bandwith), even
though in most cases, the result is not even used because the bandwidth doesn’t
match the one described by the operating class being processed.

In order to speed up the AP Capability TLV creation, avoid these useless calls whenever possible.

Closes PPM-2656

Signed-off-by: Sylvain LE ROUX <prplmesh.opensources@sagemcom.com>