controller_db : Remove duplicated byte_counter_units member variable
The member variable byte_counter_units under sBSS in Controller DB duplicates the member variable with the same name under sAgent. Consequently, one of them has to be removed. As, according to the EasyMesh spec, the Byte Counter Unit must be per Agent, not per BSS, it has been decided to remove the one under sBSS.

Closes PPM-2681

Signed-off-by: Sylvain LE ROUX <prplmesh.opensources@sagemcom.com>