bwl: whm: fix not-null check
override bool operator for AmbiorixVariant.

the AmbiorixVariant that is fed into the lambda callbacks
is constructed from a c-pointer
    AmbiorixVariant event_obj((amxc_var_t *)data, false);

AmbiorixVariant::AmbiorixVariant(amxc_var_t *data, bool is_bound) :
m_var_ctx(data), bound(is_bound)

which may be null.

the (!event_data) checks that are performed on the variable
aim to check specifically this, that the underlying c-pointer
is not null

Signed-off-by: Iacob Juc <iacob.juc_ext@softathome.com>