controller_db: dm_clear_radio_cac_capabilities: rectify data model path
Fix an issue where the data model path used in the
'dm_clear_radio_cac_capabilities()' function was incomplete. This led to
a failure in clearing old instances of the 'CACCapability.CACMethod'
object, resulting in an accumulation of instances and an abnormal memory
usage. This commit addresses the memory leakage reported in PPM-2723.

https://prplfoundationcloud.atlassian.net/browse/PPM-2723

Signed-off-by: Richard Yu <richard.yu@mitrastar.com.tw>