link_metrics_collection: check if the radio->back.iface_name is empty
Before, when this check was not present, on receiving LINK_METRIC_QUERY_MESSAGE the empty interface name was passed to a
add_wifi_neighbor function, which lead to a "Given radio iface name is empty" error message and a failed test case.
With adding this check this error is gone, and a LINK_METRIC_RESPONSE_MESSAGE is formed successfully.

PPM-2787

Signed-off-by: Denys Stolbov <d.stolbov@inango-systems.com>