slave_thread: agent_reset: ApAutoConfigurationTask should be inactive during SLAVE_INIT_DELAY_SEC
When an Agent is being reset and enters STATE_WAIT_BEFORE_INIT, the
ApAutoConfigurationTask should be inactive, but it is observed to be
active. This issue occurs because an INIT_TASK event is not being sent to
ApAutoConfigurationTask before the Agent enters STATE_WAIT_BEFORE_INIT.
As a result, ApAutoConfigurationTask remains active during
SLAVE_INIT_DELAY_SEC, which is not the intended behavior. This commit
aims to rectify the issue.

https://prplfoundationcloud.atlassian.net/browse/PPM-2835

Signed-off-by: Richard Yu <richard.yu@mitrastar.com.tw>