agent: move radio_max_bss out of the sNodeHostap struct
sNodeHostap struct is used in the

cACTION_APMANAGER_JOINED_NOTIFICATION
as well as the
cACTION_CONTROL_SLAVE_JOINED_NOTIFICATION

first message is used internally by the agent:
radio process reports info about the wifi subsystem to the main process.

second message is used between the agent and the controller.

changing the sNodeHostap struct is impacting backward compatibility
between different versions of agent and controller.

by moving the new parameter outside of the struct, directly
in the definition of the cACTION_APMANAGER_JOINED_NOTIFICATION,
removes this unwanted effect.

Signed-off-by: Iacob Juc <iacob.juc_ext@softathome.com>