db: network_map: Add status parameter in bml_conn_map
Add a code such that if an agent is inactive for
65 seconds then mark it as inactive.

65 seconds is derived from the logic of sending out
topology discovery packet by all 1905 entities
every 60 seconds and the remaining 5 seconds is
the grace time.

That status (Active/Inactive) will be displayed on
bml_conn_map which will help us to know whether an
agent is active or inactive.

PPM-2862

Signed-off-by: bsrinivasan <bsrinivasan@maxlinear.com>