controller: Fix for controller segfault
Without checking the size of list, we are trying to access the first
element of Radar stats and updating it due to this the stats_info which
is placed next to the Radar_stats list in struct sRadio, is getting
overwritten.

While accessing the stats_info segfault is seen.

Removing the timestamp update, as it is not relevant one.

Signed-off-by: rramasamy <rramasamy@maxlinear.com>